+++ module3/src/main/java/ua/com/alevel/service/OperationService.java
@@ -8,7 +8,7 @@ import java.sql.SQLException;
 public interface OperationService extends BaseService<Operation> {
     Category findCategoryByName(String name) throws SQLException;
 
-    void changeAccBalance(Integer sum,Long operationId,Boolean isIncome);
+    void changeAccBalance(Long sum,Long operationId,Boolean isIncome);
 
     void writeOutByAccId(Long id);
     void writeOutByUserId(Long id);
